#7318, move eventbus to infra
