+++ shardingsphere-jdbc/shardingsphere-jdbc-governance/src/main/java/org/apache/shardingsphere/driver/governance/internal/metadata/GovernanceSchemaChangedNotifier.java
@@ -17,7 +17,7 @@
 
 package org.apache.shardingsphere.driver.governance.internal.metadata;
 
-import org.apache.shardingsphere.governance.core.event.GovernanceEventBus;
+import org.apache.shardingsphere.infra.eventbus.ShardingSphereEventBus;
 import org.apache.shardingsphere.governance.core.event.model.schema.SchemaPersistEvent;
 import org.apache.shardingsphere.infra.metadata.schema.refresher.spi.SchemaChangedNotifier;
 import org.apache.shardingsphere.infra.metadata.schema.ShardingSphereSchema;
@@ -29,7 +29,7 @@ public final class GovernanceSchemaChangedNotifier implements SchemaChangedNotif
     
     @Override
     public void notify(final String name, final ShardingSphereSchema schema) {
-        GovernanceEventBus.getInstance().post(new SchemaPersistEvent(name, schema));
+        ShardingSphereEventBus.getInstance().post(new SchemaPersistEvent(name, schema));
     }
     
     @Override
